# backend/openapi.yaml
openapi: 3.1.0
info:
  title: WorkClub API
  version: 1.1.0
  description: |
    Multi-tenant API for agency client work, projects, tasks, time, proposals,
    invoices, teams, notifications and the restricted client portal.
  license:
    name: AGPL-3.0-only
    identifier: AGPL-3.0-only
servers:
  - url: http://localhost:8888/api
    description: Local development
...
paths:
...
  /proposal:
    get:
      tags: [Proposals]
      operationId: listProposals
      x-workclub-roles: [owner, manager]
      responses:
        '200': { $ref: '#/components/responses/Paginated' }
    post:
      tags: [Proposals]
      operationId: createProposal
      x-workclub-roles: [owner, manager]
      requestBody: { $ref: '#/components/requestBodies/JsonObject' }
      responses:
        '201': { $ref: '#/components/responses/Success' }
  /proposal/{id}:
    parameters:
      - $ref: '#/components/parameters/Id'
    get:
      tags: [Proposals]
      operationId: getProposal
      x-workclub-roles: [owner, manager]
      responses:
        '200': { $ref: '#/components/responses/Success' }
    patch:
      tags: [Proposals]
      operationId: updateProposal
      description: Enforces draft → sent and sent → accepted/declined/expired transitions.
      x-workclub-roles: [owner, manager]
      requestBody: { $ref: '#/components/requestBodies/JsonObject' }
      responses:
        '200': { $ref: '#/components/responses/Success' }
        '409': { $ref: '#/components/responses/Error' }
    delete:
      tags: [Proposals]
      operationId: archiveProposal
      x-workclub-roles: [owner, manager]
      responses:
        '200': { $ref: '#/components/responses/Success' }
        '409': { $ref: '#/components/responses/Error' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema: { $ref: '#/components/schemas/ObjectId' }
...
  requestBodies:
    JsonObject:
      required: true
      content:
        application/json:
          schema: { type: object, additionalProperties: true }
  responses:
    Success:
      description: Successful operation
      content:
        application/json:
          schema: { $ref: '#/components/schemas/SuccessEnvelope' }
    Paginated:
      description: Paginated successful operation
      content:
        application/json:
          schema:
            allOf:
              - $ref: '#/components/schemas/SuccessEnvelope'
              - type: object
                properties:
                  meta: { $ref: '#/components/schemas/Pagination' }
...
    Error:
      description: Request failed
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
  schemas:
    ObjectId:
      type: string
      pattern: '^[a-fA-F0-9]{24}$'
...
    SuccessEnvelope:
      type: object
      required: [success, result]
      properties:
        success: { const: true }
        result: {}
    ErrorEnvelope:
      type: object
      required: [success, message]
      properties:
        success: { const: false }
        message: { type: string }
        details: {}
    Pagination:
      type: object
      required: [page, limit, total]
      properties:
        page: { type: integer }
        limit: { type: integer }
        total: { type: integer }
security:
  - bearerAuth: []